+++ src/ecocjs/index.js
@@ -2,6 +2,6 @@ const ecocjs = require('bitcoinjs-lib');
 
 Object.assign(ecocjs.networks, require('./networks'));
 
-ecocjs.txs = require('./utils');
+ecocjs.utils = require('./utils');
 
 module.exports = ecocjs;
