+++ src/ecocjs/utils.js
@@ -206,7 +206,6 @@ function buildSendToContractTransaction(keyPair, contractAddress, encodedData, g
   return tx.build().toHex();
 }
 
-
 module.exports = {
   selectTxs,
   buildPubKeyHashTransaction,
